requirements.txt
mysql-connector-python==8.0.25
tencentcloud-sdk-python==3.0.398